Improve opengl mode; Add fog for EPS/SVG; Update samples; Update documentation; Replace unsigned by size_t if possible; Add proper handling braces in plot style parsing; Don't export mostly transparent primitives during 3D export.

+++ include/mgl2/addon.h
@@ -36,7 +36,7 @@ extern "C" {
 #endif
 /***************************************************************************/
 double mgl_gauss_rnd();
-void mgl_fft_freq(double *freq,unsigned nn);
+void mgl_fft_freq(double *freq,size_t nn);
 /***************************************************************************/
 /** Remove double spaces from the string */
 void mgl_strcls(char *str);
